fixed issue #30 (and also #27)

+++ src/rcpp_wt_bases_paul.cpp
@@ -24,7 +24,7 @@ const double SQRT_ONE_HALF = sqrt(.5); // NOTE: sqrt(1/2) = 1/sqrt(2)
 //' @note This c++ implementation is approx. 59% faster than the original R code
 // [[Rcpp::export]]
 List rcpp_wt_bases_paul(const NumericVector k,
-                       const int scale,
+                       const double scale,
                        const int param = -1) {
   
   const int m = param == -1 ? 4 : param;
